refactor: added type safety for componentsUrlName
