+++ src/app/types/component-info/component-min-info.ts
